Add support for sin, cos and PI
Implement support for sin, cos and PI, as per Programming Project 1, required task 3.  I've elected here to consider PI as a function
 which returns a constant value and have overloaded performOperation() accordingly.

It's nice to see that in the 'case 'PI'' call to perform operation, I just need to include the constant value in the closure, without a return statement.