Add foreground and background to themes
Programming Project 1, extra credit 1:

"Change the background and the “card back color” to match the theme. For
example, our Halloween theme has a black background and orange card
backs. Maybe a “winter” theme might have blue and white colors. A
“construction” theme could be black and yellow. UIViewController has a
property called view which is connected to the top-level view in the
scene (i.e. the view that was black in lecture)."